+++ src/References/Mapping/Event/ReferencesAdapter.php
@@ -22,7 +22,7 @@ interface ReferencesAdapter extends AdapterInterface
      * @param object        $object
      * @param bool          $single
      *
-     * @return array|string|int
+     * @return array|string|int|null array or single identifier
      */
     public function getIdentifier($om, $object, $single = true);
 
@@ -44,7 +44,7 @@ interface ReferencesAdapter extends AdapterInterface
      * @param object        $object
      * @param bool          $single
      *
-     * @return array|string|int
+     * @return array|string|int array or single identifier
      */
     public function extractIdentifier($om, $object, $single = true);
 }