+++ tests/Gedmo/Mapping/MetadataFactory/CustomDriverTest.php
@@ -1,5 +1,8 @@
 <?php
 
+namespace Gedmo\Tests\Mapping\MetadataFactory;
+
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\Mapping\ClassMetadata;
 use Doctrine\Persistence\Mapping\Driver\MappingDriver;
 use Gedmo\Tests\Mapping\Fixture\Unmapped\Timestampable;
@@ -15,6 +18,16 @@ use Gedmo\Tests\Mapping\Fixture\Unmapped\Timestampable;
  */
 final class CustomDriverTest extends \PHPUnit\Framework\TestCase
 {
+    /**
+     * @var Timestampable
+     */
+    private $timestampable;
+
+    /**
+     * @var EntityManagerInterface
+     */
+    private $em;
+
     protected function setUp(): void
     {
         $config = new \Doctrine\ORM\Configuration();