+++ tests/Gedmo/Tree/Fixture/Closure/Category.php
@@ -2,6 +2,8 @@
 
 namespace Gedmo\Tests\Tree\Fixture\Closure;
 
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 
@@ -37,6 +39,16 @@ class Category
      */
     private $parent;
 
+    /**
+     * @var Collection<int, CategoryClosure>
+     */
+    private $closures;
+
+    public function __construct()
+    {
+        $this->closures = new ArrayCollection();
+    }
+
     public function getId()
     {
         return $this->id;
@@ -52,7 +64,7 @@ class Category
         return $this->title;
     }
 
-    public function setParent(Category $parent = null)
+    public function setParent(self $parent = null)
     {
         $this->parent = $parent;
     }