+++ tests/Gedmo/Tree/Fixture/UserLDAP.php
@@ -10,8 +10,8 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class UserLDAP extends User
 {
-    public function __construct($ldapUserName)
+    public function __construct(string $ldapUserName = 'next@something.com')
     {
-        parent::__construct('next@something.com', 'pass');
+        parent::__construct($ldapUserName, 'pass');
     }
 }